+++ integration/graphql_types_test.go
@@ -131,8 +131,8 @@ type Party struct {
 	// TODO:
 	// Positions []Position
 	// Margins []MarginLevels
-	// Withdrawals []WithDrawal
-	Deposits []Deposit
+	Withdrawals []Withdrawal
+	Deposits    []Deposit
 	// Delegations []Delegation
 	// Stake PartyStake
 	// Rewards []Reward
@@ -270,4 +270,17 @@ type Deposit struct {
 type NetworkParameter struct {
 	Key   string
 	Value string
+}
+
+type Withdrawal struct {
+	ID                 string
+	Party              Party
+	Amount             string
+	Asset              Asset
+	Status             string
+	Ref                string
+	Expiry             string
+	TxHash             string
+	CreatedTimestamp   string
+	WithdrawnTimeStamp string
 }