+++ integration/parties_test.go
@@ -5,6 +5,7 @@ import "testing"
 func TestParties(t *testing.T) {
 	queries := map[string]string{
 		"Deposits":      "{ parties { deposits{ id, party { id }, amount, asset { id }, status, createdTimestamp, creditedTimestamp, txHash } } }",
+		"Withdrawals":   "{ parties { withdrawals { id, party { id }, amount, asset { id }, status, ref, expiry, txHash, createdTimestamp, withdrawnTimestamp } } }",
 		"Delegations":   "{ parties{ id delegations{ node { id }, party{ id }, epoch, amount } } }",
 		"Proposals":     "{ parties{ id proposals{ id votes{ yes { totalNumber } no { totalNumber } } } } }",
 		"Votes":         "{ parties{ id votes{ proposalId vote{ value } } } }",